requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
blis==0.7.11
bs4==0.0.2
catalogue==2.0.10
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.5
configobj==5.0.8
configparser==7.0.0
cymem==2.0.8
dataclasses-json==0.6.7
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
etelemetry==0.3.1
filelock==3.14.0
fitz==0.0.1.dev2
Flask==3.0.0
Flask-CLI==0.4.0
Flask-Cors==4.0.1
frozenlist==1.4.1
fsspec==2024.6.0
greenlet==3.0.3
h11==0.14.0
haversine==2.8.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.6
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.5
langchain-core==0.2.9
langchain-text-splitters==0.2.1
langcodes==3.4.0
langsmith==0.1.81
language_data==1.2.0
llama-index==0.10.45
llama-index-agent-openai==0.2.7
llama-index-cli==0.1.12
llama-index-core==0.10.44
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.22
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.25
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.4
llamaindex-py-client==0.1.19
looseversion==1.3.0
lxml==5.2.2
marisa-trie==1.1.1
MarkupSafe==2.1.3
marshmallow==3.21.3
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nibabel==5.2.1
nipype==1.8.6
nltk==3.8.1
numpy==1.26.4
openai==1.30.5
orjson==3.10.5
packaging==24.0
pandas==2.2.2
pathlib==1.0.1
pillow==10.3.0
preshed==3.0.9
prov==2.0.0
pydantic==2.7.2
pydantic_core==2.18.3
pydot==2.0.0
pyparsing==3.1.2
pypdf==4.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==7.0.0
regex==2024.5.15
requests==2.31.0
scipy==1.13.1
setuptools==70.0.0
simplejson==3.19.2
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.30
srsly==2.4.8
striprtf==0.0.26
tenacity==8.3.0
thinc==8.2.3
tiktoken==0.7.0
tqdm==4.66.4
traits==6.3.2
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.0
wasabi==1.1.3
weasel==0.3.4
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
